Adds botany jumpskirt and some of the framework for the new additions system (#4)
* Added modular vending machine and minimal clothing support

Vending machines needed a little more code to make the russ_products and such work as intended. I simply copy pasta'd hippies code and changed references to hippie to russ. Clothing support just makes it so that things in the jumpsuit slot will get the icons from the right spot if they are made according to the modular system.

* Adds Botany Jumpskirt

Just a recolor of the secskirt that already existed, should be available in the department vending machines as well.

* Update RussStation.dme

need to include things.

* Revert "Update RussStation.dme"

This reverts commit fe0802963d91954ee2bbe08492555b448604b42d.

* Botany jumpskirt and modular system frameworks

New botany jumpskirt that should be available in the department wardrobe. As well as a few of the dependencies copied from hippie to make it work.

* Remove useless comments