Pubbypipe (#32)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull request process. -->

> ## About The Pull Request
In this pull request, one of the most inconvenient things about being a Janitor on Pubby Station is addressed: A malfunctioning disposal system. As part of the changes, a single disposal pipe was removed, and replaced with a more simplified correct facing pipe. In doing this, the disposal system once again works as it was always intended- nay. As God intended for it to work: Correctly.

Also, the Camdola code is moved from above the cable code to below the cable code. This was done because Byond is a fickle mistress and decided that the code was better moved, even though it has absolutely no bearing whatso ever on the game itself.

## Why It's Good For The Game
These changes are good for the game because they fix the longstanding problem of not being able to flush a body down disposals and have it bless Cargo with miasma. Or flush a monkey down disposals to bless Cargo with monkeys. Or flush various used Syndicate items down Cargo to attempt to hide evidence but fail because there's a detective on shift who somewhat knows how to do their job. And many more.

The Camdola change is good for the game because it pleases the Byond gods and ensures that we get at least another week or so without any major issues. Failure to ensure the Camdola changes are pushed can result in unstable, potentially unusable, flammable, inflammable, and otherwise inoperable code.

## Changelog
:cl:
fix: Fixes the disposal system on Pubby Station.
/:cl:


* Repairs the disposal loop on Pubby

* Adds the changelog